+++ app/src/main/java/com/github/brugapp/brug/di/sign_in/module/DatabaseAuthModule.kt
@@ -1,5 +1,7 @@
-package com.github.brugapp.brug.di.sign_in
+package com.github.brugapp.brug.di.sign_in.module
 
+import com.github.brugapp.brug.di.sign_in.AuthDatabase
+import com.github.brugapp.brug.di.sign_in.firebase.AuthFirebase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
