+++ app/src/main/java/com/github/brugapp/brug/di/sign_in/module/SignInResultHandlerModule.kt
@@ -1,5 +1,7 @@
-package com.github.brugapp.brug.di.sign_in
+package com.github.brugapp.brug.di.sign_in.module
 
+import com.github.brugapp.brug.di.sign_in.SignInResultHandler
+import com.github.brugapp.brug.di.sign_in.google.SignInResultHandlerGoogle
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
